scripts/python/requirements.txt
Faker==22.6.0
locust==2.29.0
Requests==2.32.3
websocket_client==1.7.0
